Revert "ec2: add new command for remote connect (#3649)"
This reverts commit 4c9713be84882cdad8728f364b02f8546a202b72.
