+++ package.json
@@ -1111,11 +1111,7 @@
                     "command": "aws.auth.manageConnections"
                 },
                 {
-                    "command": "aws.ec2.openRemoteConnection",
-                    "when": "aws.isDevMode"
-                },
-                {
-                    "command": "aws.ec2.openTerminal",
+                    "command": "aws.ec2.connectToInstance",
                     "when": "aws.isDevMode"
                 },
                 {
@@ -1313,7 +1309,7 @@
                     "group": "inline@1"
                 },
                 {
-                    "command": "aws.ec2.openTerminal",
+                    "command": "aws.ec2.connectToInstance",
                     "group": "0@1",
                     "when": "viewItem == awsEc2Node"
                 },
@@ -2060,18 +2056,8 @@
                 }
             },
             {
-                "command": "aws.ec2.openTerminal",
-                "title": "%AWS.command.ec2.openTerminal%",
-                "category": "%AWS.title%",
-                "cloud9": {
-                    "cn": {
-                        "category": "%AWS.title.cn%"
-                    }
-                }
-            },
-            {
-                "command": "aws.ec2.openRemoteConnection",
-                "title": "%AWS.command.ec2.openRemoteConnection%",
+                "command": "aws.ec2.connectToInstance",
+                "title": "%AWS.command.ec2.connectToInstance%",
                 "category": "%AWS.title%",
                 "cloud9": {
                     "cn": {
