+++ package.nls.json
@@ -108,8 +108,7 @@
     "AWS.command.refreshAwsExplorer": "Refresh Explorer",
     "AWS.command.refreshCdkExplorer": "Refresh CDK Explorer",
     "AWS.command.cdk.help": "View CDK Documentation",
-    "AWS.command.ec2.openTerminal": "Open terminal to EC2 instance...",
-    "AWS.command.ec2.openRemoteConnection": "Connect to EC2 instance in New Window...",
+    "AWS.command.ec2.connectToInstance": "Connect to EC2 Instance...",
     "AWS.command.ecr.copyTagUri": "Copy Tag URI",
     "AWS.command.ecr.copyRepositoryUri": "Copy Repository URI",
     "AWS.command.ecr.createRepository": "Create Repository...",
